chore: migrate spellchecker-spec.ts to vitest
